refactor: update show offers and filter modal
