+++ webapp/package.json
@@ -35,6 +35,7 @@
     "prop-types": "^15.7.2",
     "qrcode.react": "^1.0.0",
     "react": "^16.13.1",
+    "react-cookie": "^4.0.3",
     "react-dom": "^16.13.1",
     "react-facebook": "^8.1.4",
     "react-google-login": "^5.1.21",
