+++ webapp/src/routes/Home/Home.js
@@ -1,4 +1,5 @@
-import React from 'react'
+import React, { useEffect, useState } from 'react'
+import { useQuery } from '@apollo/react-hooks'
 import PropTypes from 'prop-types';
 import { makeStyles, useTheme } from '@material-ui/styles'
 import useMediaQuery from '@material-ui/core/useMediaQuery'
@@ -14,15 +15,15 @@ import Toolbar from '@material-ui/core/Toolbar';
 import SearchIcon from '@material-ui/icons/Search';
 import StarIcon from '@material-ui/icons/Star';
 import FavoriteIcon from '@material-ui/icons/Favorite';
-
-
 import MapModal from '../../components/MapModal'
 import FilterHome from '../../components/FilterHome'
-import Offers from '../Offers'
+import ShowOffers from './ShowOffers';
 import Banks from '../Banks';
 import Sponsors from '../Sponsors';
 import mobileBgImage from '../../assets/the-world.png'
 import bgImage from '../../assets/lifebank-hero-bg.png'
+
+import { GET_OFFERS_QUERY, GET_LOCATIONS_QUERY } from '../../gql'
 
 const useStyles = makeStyles((theme) => ({
   homeHeader: {
@@ -112,34 +113,88 @@ function a11yProps(index) {
   };
 }
 
-
-
 const Home = () => {
   const classes = useStyles()
   const [currentUser] = useUser()
-  const [value, setValue] = React.useState(0);
+  const [value, setValue] = useState(0);
   const theme = useTheme()
   const isDesktop = useMediaQuery(theme.breakpoints.up('md'), {
     defaultMatches: true
   })
 
-  const handleChange = (event, newValue) => {
+  const [loadingOffers, setLoadingOffers] = useState(true);
+  const [offers, setOffers] = useState([]);
+  const [loadingLifebanks, setLoadingLifebanks] = useState(true);
+  const [lifebanks, setLifebanks] = useState([]);
+  const [loadingSponsors, setLoadingSponsors] = useState(true);
+  const [sponsors, setSponsors] = useState([]);
+  const [openDialogFilter, setOpenDialogFilter] = useState(false)
+
+  const { refetch: getAllOffers } = useQuery(GET_OFFERS_QUERY, { active: true }, { skip: true })
+  const { refetch: getAllBanks } = useQuery(GET_LOCATIONS_QUERY, {}, { skip: true })
+  const { refetch: getAllSponsors } = useQuery(GET_LOCATIONS_QUERY, {}, { skip: true })
+
+  const getOffers = async () => {
+    const { data } = await getAllOffers({ active: true })
+    const dataTemp = data.offer
+
+    setOffers(dataTemp)
+    setLoadingOffers(false)
+  }
+
+  const getLifebanks = async () => {
+    const { data } = await getAllBanks({})
+    let dataTemp = data.location
+    dataTemp = dataTemp.filter(bank => bank.type === "LIFE_BANK")
+
+    setLifebanks(dataTemp)
+    setLoadingLifebanks(false)
+  }
+
+  const getSponsors = async () => {
+    const { data } = await getAllSponsors({})
+    let dataTemp = data.location
+    dataTemp = dataTemp.filter(bank => bank.type === "SPONSOR")
+
+    setSponsors(dataTemp)
+    setLoadingSponsors(false)
+  }
+
+  useEffect(() => {
+    getOffers()
+  }, [getAllOffers])
+
+  useEffect(() => {
+    getLifebanks()
+  }, [getAllBanks])
+
+  useEffect(() => {
+    getSponsors()
+  }, [getAllSponsors])
+
+  const handlerApplyFilterOffer = (data) => {
+    setOffers(data)
+  }
+
+  const handleChangeTabs = (event, newValue) => {
     setValue(newValue);
   };
-
 
   const HomeMobile = () => {
     return (
       <>
         <AppBar position="static" className={classes.appBarTab}>
-          <Tabs className={classes.tabs} value={value} onChange={handleChange} aria-label="Home tabs" centered variant="fullWidth">
+          <Tabs className={classes.tabs} value={value} onChange={handleChangeTabs} aria-label="Home tabs" centered variant="fullWidth">
             <Tab label="OFFERS" {...a11yProps(0)} />
             <Tab label="BANKS" {...a11yProps(1)} />
             <Tab label="SPONSORS" {...a11yProps(2)} />
           </Tabs>
           <TabPanel value={value} index={0} className={classes.tabPanel}>
             <Paper style={{ height: 'calc(100vh - 128px)', overflow: 'auto', border: "none" }}>
-              <Offers />
+              <ShowOffers
+                offers={offers}
+                loading={loadingOffers}
+              />
             </Paper>
           </TabPanel>
           <TabPanel value={value} index={1}>
@@ -159,7 +214,11 @@ const Home = () => {
               <SearchIcon className={classes.iconBottomAppBar} />
             </IconButton>
             <MapModal />
-            <FilterHome />
+            <FilterHome
+              handlerApplyFilterOffer={handlerApplyFilterOffer}
+              openDialogFilter={openDialogFilter}
+              setOpenDialogFilter={setOpenDialogFilter}
+            />
             <IconButton disabled>
               <StarIcon className={classes.iconBottomAppBar} />
             </IconButton>
