--- webapp/src/routes/Offers/index.js
@@ -1,1 +1,0 @@
-export { default } from './Offers'
