+++ webapp/src/routes/index.js
@@ -8,7 +8,6 @@ import About from './About'
 import TermsOfUse from './TermsOfUse'
 import Help from './Help'
 import OffersManagement from './OffersManagement'
-import Offers from './Offers'
 import OfferPage from './OfferPage'
 import Home from './Home'
 import EmailVerification from './EmailVerification'
@@ -45,11 +44,6 @@ export default [
     name: 'donation',
     path: '/donations',
     component: Donation
-  },
-  {
-    name: 'offers',
-    path: '/offers',
-    component: Offers
   },
   {
     name: 'offer-page',
